+++ MQTTClient/MQTTClientTests/SwiftTests.swift
@@ -87,7 +87,7 @@ class SwiftTests : XCTestCase, MQTTSessionDelegate {
         sessionReceived = true;
     }
     
-    func subAckReceived(session: MQTTSession!, msgID: UInt16, grantedQoss qoss: [AnyObject]!) {
+    func subAckReceived(session: MQTTSession!, msgID: UInt16, grantedQoss qoss: [NSNumber]!) {
         sessionSubAcked = true;
     }